refactor: move updateAchievements to its own file
